+++ src/App.css
@@ -6,10 +6,13 @@
   padding: 0;
 }
 
-.App {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  /* height: 100vh; */
+body,
+html {
+  height: 100%;
+}
+
+#root {
+  position: relative;
+  height: 100%;
+  padding: 5px;
 }
