+++ src/components/WeatherRight/WeatherRight.module.scss
@@ -1,3 +1,5 @@
+@import '../../responsive.scss';
+
 .weather-right {
   width: 45%;
   padding: var(--padding-3) var(--padding-4);
@@ -9,4 +11,7 @@
     rgba(29, 29, 185, 1) 100%,
     rgba(29, 29, 185, 0.6617997540813201) 100%
   );
+  @include mdScreen {
+    width: 50%;
+  }
 }
