+++ src/pages/Main/index.js
@@ -1,10 +1,10 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import './main.scss';
 import WeatherLeft from '@components/WeatherLeft';
 import WeatherRight from '@components/WeatherRight';
 import PopUp from '@components/PopUp';
 import weatherApi from '@api/weatherApi';
-import { debounce } from 'lodash';
+import useForecast from '../Main/hooks/useForecast';
 import { changeSpeedUnit, getLocalDay, getLocalMonth, getLocalTime } from '@utility/formatData';
 
 import IconWind from '@animated/dust-wind.svg';
@@ -15,114 +15,20 @@ import IconSunRise from '@animated/clear-day.svg';
 import IconSunSet from '@animated/haze-day.svg';
 
 export default function MainPage() {
-  const [forecastData, setForecastData] = useState({});
-  const [searchText, setSearchText] = useState('Hanoi');
-  const [units, setUnits] = useState('metric');
+  // const [forecastData, setForecastData] = useState({});
   const [loading, setLoading] = useState(true);
   const [popUpError, setPopUpError] = useState(false);
-  const searchRef = useRef(null);
+  const [searchText, setSearchText] = useState('');
+  const [units, setUnits] = useState('metric');
 
-  // for default display when first load
-  useEffect(() => {
-    fetchWeatherData(searchText);
-  }, [units, searchRef.current]);
-
-  const fetchWeatherData = async (searchText) => {
-    try {
-      setLoading(true);
-      const responseLocation = await weatherApi.getWeatherData('weather', {
-        q: searchText,
-        appid: process.env.REACT_APP_API_KEY,
-      });
-      const { lat, lon } = responseLocation.coord;
-      const responseForecast = await weatherApi.getWeatherData('onecall', {
-        lat,
-        lon,
-        appid: process.env.REACT_APP_API_KEY,
-        units,
-      });
-
-      // time info
-      const { current, timezone } = responseForecast;
-      let localDay = getLocalDay(current.dt, timezone);
-      let localMonth = getLocalMonth(current.dt, timezone);
-      let localTime = getLocalTime(current.dt, timezone);
-      let localName = timezone.replace('/', ', ');
-
-      // weather info
-      let { sunrise, sunset, temp, pressure, humidity, uvi, wind_speed } = current;
-      let { description, icon } = current.weather[0];
-      const iconScr = weatherApi.getWeatherIconScr(icon);
-      description = description.replace(/\w\S*/g, (w) => w.replace(/^\w/, (c) => c.toUpperCase()));
-      sunrise = getLocalTime(sunrise, timezone);
-      sunset = getLocalTime(sunset, timezone);
-
-      if (units === 'imperial') {
-        wind_speed = changeSpeedUnit(wind_speed);
-        wind_speed = Math.round(wind_speed * 100) / 100;
-        temp = `${temp}°F`;
-      } else {
-        temp = `${temp}°C`;
-      }
-
-      // hourly temperature
-      let hourlyTemp = [];
-      for (let hour = 0; hour <= 24; hour += 3) {
-        if (hourlyTemp.length < 9) {
-          let temp = responseForecast.hourly[hour].temp;
-          if (units === 'imperial') {
-            temp = `${temp}°F`;
-          } else {
-            temp = `${temp}°C`;
-          }
-          hourlyTemp.push(temp);
-        }
-      }
-
-      //daily rain chance
-      const dailyRainChance = responseForecast.daily.slice(0, 7).map((day) => {
-        return Math.round(day.pop * 100);
-      });
-
-      setForecastData({
-        sunrise,
-        sunset,
-        temp,
-        pressure,
-        humidity,
-        uvi,
-        wind_speed,
-        description,
-        iconScr,
-        localDay,
-        localMonth,
-        localTime,
-        localName,
-        hourlyTemp,
-        dailyRainChance,
-      });
-      debounceLoading();
-    } catch (err) {
-      if (err?.response?.status === 404) {
-        setPopUpError(true);
-        setLoading(false);
-      } else {
-        console.log('Failed to fetch weather data', err);
-      }
-    }
-  };
-
-  const debounceLoading = debounce(() => {
-    setLoading(false);
-  }, 1000);
-
-  const handleSearchSubmit = (e) => {
-    e.preventDefault();
-    if (popUpError) {
-      return;
-    }
-    fetchWeatherData(searchText);
-  };
+  const { forecastData, handleSearchSubmit } = useForecast(
+    searchText,
+    units,
+    loading,
+    setLoading,
+    popUpError,
+    setPopUpError,
+  );
 
   const handleSearchChange = (e) => {
     e.preventDefault();
@@ -131,11 +37,8 @@ export default function MainPage() {
 
   const handleLocationClick = (e) => {
     if (searchText === 'Hanoi') return;
-    searchRef.current.value = 'Hanoi';
-    setSearchText(searchRef.current.value);
-    // fetchWeatherData(searchText);
+    setSearchText('Hanoi');
   };
-  console.log(searchText);
   const weatherInfoList = [
     { id: 1, name: 'Wind Speed', iconSrc: IconWind, value: `${forecastData.wind_speed} m/s` },
     { id: 2, name: 'Humidity', iconSrc: IconCloudRain, value: `${forecastData.humidity} %` },
@@ -161,7 +64,7 @@ export default function MainPage() {
         onSearchSubmit={handleSearchSubmit}
         onSearchChange={handleSearchChange}
         onLocationClick={handleLocationClick}
-        searchRef={searchRef}
+        value={searchText}
       />
       <WeatherRight forecastData={forecastData} units={units} loading={loading} sunInfoList={sunInfoList} />
     </section>
