+++ src/responsive.scss
@@ -1,0 +1,28 @@
+$XS: 480px;
+$SM: 768px;
+$MD: 1024px;
+$XL: 1400px;
+
+@mixin xsScreen {
+  @media (max-width: #{$XS - 1px}) {
+    @content;
+  }
+}
+
+@mixin smScreen {
+  @media (min-width: #{$XS}) and (max-width: #{$MD - 1px}) {
+    @content;
+  }
+}
+
+@mixin mdScreen {
+  @media (min-width: #{$MD}) and (max-width: #{$XL - 1px}) {
+    @content;
+  }
+}
+
+@mixin xlScreen {
+  @media (min-width: #{$XL}) {
+    @content;
+  }
+}
